[runtime] Rename BaseDescriptors to DescriptorRegistry, move to own file
