+++ src/js/runtime/bytecode/function.rs
@@ -280,7 +280,7 @@ impl BytecodeFunction {
         let size = Self::calculate_size_in_bytes(bytecode.len());
         let mut object = cx.alloc_uninit_with_size::<BytecodeFunction>(size)?;
 
-        set_uninit!(object.descriptor, cx.base_descriptors.get(HeapItemKind::BytecodeFunction));
+        set_uninit!(object.descriptor, cx.descriptors.get(HeapItemKind::BytecodeFunction));
         set_uninit!(object.constant_table, constant_table.map(|c| *c));
         set_uninit!(object.exception_handlers, exception_handlers.map(|h| *h));
         set_uninit!(object.realm, *realm);
@@ -322,7 +322,7 @@ impl BytecodeFunction {
             new_target_index = Some(0);
         }
 
-        set_uninit!(object.descriptor, cx.base_descriptors.get(HeapItemKind::BytecodeFunction));
+        set_uninit!(object.descriptor, cx.descriptors.get(HeapItemKind::BytecodeFunction));
         set_uninit!(object.constant_table, None);
         set_uninit!(object.exception_handlers, None);
         set_uninit!(object.realm, *realm);
