+++ src/js/runtime/string_value.rs
@@ -759,7 +759,7 @@ impl FlatString {
         let size = Self::calculate_size_in_bytes(len, StringWidth::OneByte);
         let mut string = cx.alloc_uninit_with_size::<FlatString>(size)?;
 
-        set_uninit!(string.descriptor, cx.base_descriptors.get(HeapItemKind::String));
+        set_uninit!(string.descriptor, cx.descriptors.get(HeapItemKind::String));
         set_uninit!(string.len, len);
         set_uninit!(string.kind, StringKind::OneByte);
         set_uninit!(string.is_interned, false);
@@ -782,7 +782,7 @@ impl FlatString {
         let size = Self::calculate_size_in_bytes(len, StringWidth::TwoByte);
         let mut string = cx.alloc_uninit_with_size::<FlatString>(size)?;
 
-        set_uninit!(string.descriptor, cx.base_descriptors.get(HeapItemKind::String));
+        set_uninit!(string.descriptor, cx.descriptors.get(HeapItemKind::String));
         set_uninit!(string.len, len);
         set_uninit!(string.kind, StringKind::TwoByte);
         set_uninit!(string.is_interned, false);
@@ -1298,7 +1298,7 @@ impl ConcatString {
     ) -> EvalResult<Handle<StringValue>> {
         let mut string = cx.alloc_uninit::<ConcatString>()?;
 
-        set_uninit!(string.descriptor, cx.base_descriptors.get(HeapItemKind::String));
+        set_uninit!(string.descriptor, cx.descriptors.get(HeapItemKind::String));
         set_uninit!(string.len, len);
         set_uninit!(string.kind, StringKind::Concat);
         set_uninit!(string.width, width);
